Added threadpool and some optimizations
The Roots type now has a threadpool for committing transactions, in
theory allowing for transactions that hit multiple trees to be better
parallelized.

The tree modifications while in a transaction now happen in memory,
except for new values being written to disk. There is a new task that is
used to write the changes to disk. This requires additional memory while
a transaction is being executed -- but it also means less
allocations/deallocations if multiple edits are made to the same tree in
the same transaction.

The rest of the changes are focused on a few simple optimizations:

- Buffer reuse -- less Vec::new()'s
- Minimize copies -- Some to_vecs were removed