Bug fixes with transaction handling.
Through tracking down the ultimate bug I was looking for (the one-line
change in btree_entry.rs), I noticed some potential edge cases in how
the transaction trees were being committed. Under the current
implementation of the file manager, they couldn't be exploited, but to
make it less likely to happen in the future, I refactored how the tree
locks flow to be more explicit when the transaction releases them.