Enabling multiple Roots instances to share state
One of the needs of BonsaiDb ended up being the ability to have a
per-database transaction log. This means it makes more sense for Roots
to be instantiated at the database level, not the server level. This is
how Sled used to be utilized within BonsaiDb, but it didn't provide a
mechanism to share background tasks or cache limits -- not ideal for
building a server.

However, Nebari can provide those APIs and that's what this does --
enables sharing thread pools and file managers, which allows BonsaiDb
server to create multiple Roots instances, one for each database, and
share a single thread pool and set of caches.