Implemented tree compaction
Closes #3.

This implementation is a little rough -- it currently will load the
entire set of nodes into memory as part of the process of copying. For
large databases, that could be a lot of memory.

I think a way to evolve this would be to pass a parameter into
copy_data_to that indictes whether `self` should serialize automatically
or not.

The compact process purposefluly avoids using the ChunkCache and it
re-adds CRC checking. Since most databases will run compaction
periodically, this seems like a good balance for now (refs #9).