Initial work on adding compaction to Roots
This is still a work in progress. Currently, the locking strategy for
compaction is much too aggressive, and it's due to me having trouble
narrowing down a bug over the last two days that ultimately ended up
being caused by the unit test itself having a bug in it.

What this commit does:

- Introduces backtraces for errors. I got tired of not having backtraces
  of where the errors originate. Currently, errors in nebari are for
  true error situations, not flow control, so the overhead of generating
  the backtrace resolved seems fine.
- Adds `Tree::compact()`, which executes the underlying compact
  operation on TreeFile.
- Adds a unit test designed to stress-test the concurrency model of
  Roots: multiple workers doing a lot of transactions while another
  thread continuously compacts the database.

This is working for unversioned trees, but not versioned trees. Last
night, I realized I had implemented the By Sequence index in a
completely wrong way -- currently the keys in the By Sequence index are
the document IDs, not the sequence IDs. Since we're not really using the
by sequence index yet, I hadn't really noticed the issue. Before
refactoring that, I wanted to commit what I had done.

Due to the locking currently employed, the DatabaseCompacted error never
actually gets raised. Before this is merged, either that error should go
away or we'll re-introduce the race condition that allows it to happen
-- in which case we'll need to handle that error internally in Roots and
retry operations if that error is raised.