VersionedTreeRoot is now implemented correctly
Both roots now primarily modify the by_id_root, and the by_sequence_root
now uses sequence IDs as the key, allowing for a full databse scan from
the initial write (that still exists in the file) through all writes in
the file.

For the roots compaction test, I've tried throwing a massive number of
threads at it to break it. I contend there is a race condition that can
cause the DatabaseCompacted error to occur between the file being opened
and when the state is locked for read. As far as I can reason, it's
possible for the database compaction process to replace files between
those moments in time, causing the file_id on the state to no longer
match the ID of the file opened from the manager. The window is small,
but try as I might I couldn't get the error to arise.

So, we still need to handle this edge case -- but I think with the
current aggressive locking strategy, it can only happen on read
operations.