Implemented tree rebalancing
Fixes #6

This implementation restricts leaf nodes to contain no fewer than the
current b-tree "order" / 2 - 1. As a tree grows, the order grows
dynamically, so nodes that haven't been modified might be absorbed the
next time they are touched.

The compact process should probably also clean up nodes, so that a
compact tidies up those nodes.

The other limitation currently is that interior nodes have no checks for
minimum children. That was mostly to simplify implementation, but we
should definitely do that in the future.