Delay resolving error backtraces
@ModProg reported an issue when experimenting with inserting very large
transactions in BonsaiDb. When inserting too much data in a transaction,
he was able to get into a state where it seemed like BonsaiDb had
frozen.

I added a test to simulate the underlying behavior, as I thought all the
error cases had been handled. It turns out, Nebari was behaving
correctly. The "frozen" behavior was becuase Nebari was trying to find a
valid page start, but had to scan the entire file page by page looking
for it.

This is due to how writes happen in Roots -- data is written to the file
before the transaction succeeds, but a new header is only written once
the transaction is being committed. BonsaiDb wrote all the contents of
the collection to the tree, and Nebari threw an error when BonsaiDb
attempted to set the associated data for the transaction log.

Upon next opening of the database, Nebari tried to scan the tree file
for where it should begin. This process is normally fast, but with so
much junk data at the end of the file, it was fairly slow.

I noticed a lot of time was being spent in calculating error backtraces,
so I moved the backtrace into a Mutex so that we can lock it and resolve
when we go to print the error. This makes recovery in this failure
scenario much faster, even though it still ends up scanning the entire
file.