Reimplement MemoryFile locking strategy
The original goal of this was to unit test compact under Roots using
MemoryFiles. This uncovered a deadlock unique to MemoryFile.

This refactor ensures that readers can't block writers. Previously the
need to access certain information behind a shared lock caused issues,
as readers and writers could block. This isn't actually an issue
normally, except the operations necessary to compact required access to
some of the information while the file was locked due to the
requirements of FileOp.

By returning a unique copy of the metadata for each OpenMemoryFile
instance, it solves all of the locking issues. and operates as real file
handles would -- each handle has a unique position tracked within the
file.