Fixing transaction log test
Was reviewing coverage and saw a block of test code that never invoked. Turns out a good chunk of testing was being skipped.

The testing encryption vault now adds a header to be able to detect an unencrypted payload. I've added #35 to give us more validation inside of transaction log scanning, but for now, the test has been flipped to try using a vault on an unencrypted payload instead of the other direction. Functionally the same, but the CRCs would have allowed it to work either way.