TransactionManager now ensures log entry order
While working on other things, I ran into a random failure in a test I
was running, and I realized that it stemmed from guarantees that the
transaction log was assuming about the order of the file. Part of me
didn't like the idea of slowing the transaction log down a bit by
forcing it to wait for transactions if they were committed out of order.
However, without significantly changing how the transaction log works,
this was a guarantee that needed to be made.

TransactionLog now checks for the order and returns an error when an
attempt is made.

These changes ensure that scan() can always return every transaction log
entry.