Removing TransactionLog::current_transaction_id
In reviewing other code that was calling current_transaction_id, I
realized there was a naming issue with this function. Since it was a
simple pass-through function, rather than just fixing its code, I opted
to keep that code centralized on State.

This makes it a true breaking change for anyone who was using it, but it
forces them to decide whether they want the last written id or the next
id as part of the upgrade, rather than me silently causing
incompatibility issues.