Added better suppport for working with indexes
While working on khonsulabs/bonsaidb#225, I had to do major updates on
how documents are stored. In the new storage scheme, the key in Nebari
is the DocumentId, Revision::id is Nebari's SequenceId, and
Revision::hash is indexed and stored as an embedded index. This means
that retrieving a document needs access to the index to fully construct
the Document record -- hence these new APIs.

Refs: khonsulabs/bonsaidb#250