Allow some transaction concurrency
This introduces a third ActiveState that is managed on a
per-committing-transaction basis. Rather than wait to release the locks
on trees until after the transaction is fully synchronized, these
changes release the locks after the file is fsynced but before the
transaction log is synced.

Before the transaction log is synced and the locks are released, the
write state is cloned. After the transaction is confirmed, the state is
published (taking care to ensure an older state doesn't overwrite a
newer state since the OS controls waking threads).

This works, but doesn't really offer parallelism due to the inability to
batch the fsync operations. To complete this work, I need to figure out
a good way to let the transaction manager issue the fsync commands, but
be smart enough to only do it once per tree per batched commit.