Move tree syncing and publishing to tx manager
The limiting factor of the last commit was that each write operation on
each tree still had its own fsync operation. This change moves the
responsibility of fsyncing to the transaction manager, which supports
batching multiple transactions together already.

Now the transaction manager keeps a unique list of paths that need to be
synchronized as part of the batch. This allows multiple pending
transactions for a single tree to pile up while waiting on the current
batch to be fsynced.

This is safe because when opening a file with a transaction manager, the
file is automatically rolled back to the last confirmed transaction id.
The transaction manager ensures each tree is fsynced before writing the
transaction log entry and fsyncing the transaction. All new data written
by pending transactions will be rolled back if a power loss occurs while
the next batch is being fsynced unless the transaction log is fully
synced.

These changes allowed a significant speedup to BonsaiDb's commerce
benchmark, and the new bottleneck is the reliance on two fsyncs.